@@ -1,68 +1,59 @@
-import React, { Component } from 'react';
-import { ReactiveBase, DataSearch } from '@appbaseio/reactivesearch';
+import React, { useState } from "react";
+import { ReactiveBase, DataSearch } from "@appbaseio/reactivesearch";
 
-import Header from './components/Header';
-import Results from './components/Results';
+import Header from "./components/Header";
+import Results from "./components/Results";
 
-import theme from './theme';
-import './App.css';
+import theme from "./theme";
+import "./App.css";
 
-class App extends Component {
-	constructor(props) {
-		super(props);
-		this.state = {
-			currentTopics: [],
-		};
-	}
+const App = props => {
+  const [currentTopics, setTopics] = useState([]);
 
-	setTopics = (currentTopics) => {
-		this.setState({
-			currentTopics: currentTopics || [],
-		});
-	}
+  const toggleTopic = topic => {
+    const nextState = currentTopics.includes(topic)
+      ? currentTopics.filter(item => item !== topic)
+      : currentTopics.concat(topic);
+    setTopics(nextState);
+  };
 
-	toggleTopic = (topic) => {
-		const { currentTopics } = this.state;
-		const nextState = currentTopics.includes(topic)
-			? currentTopics.filter(item => item !== topic)
-			: currentTopics.concat(topic);
-		this.setState({
-			currentTopics: nextState,
-		});
-	}
-
-	render() {
-		return (
-			<section className="container">
-				<ReactiveBase
-					app="gitxplore-app"
-					credentials="4oaS4Srzi:f6966181-1eb4-443c-8e0e-b7f38e7bc316"
-					type="gitxplore-latest"
-					theme={theme}
-				>
-					<div className="flex row-reverse app-container">
-						<Header currentTopics={this.state.currentTopics} setTopics={this.setTopics} />
-						<div className="results-container">
-							<DataSearch
-								componentId="repo"
-								filterLabel="Search"
-								dataField={['name', 'description', 'name.raw', 'fullname', 'owner', 'topics']}
-								placeholder="Search Repos"
-								iconPosition="left"
-								autosuggest={false}
-								URLParams
-								className="data-search-container results-container"
-								innerClass={{
-									input: 'search-input',
-								}}
-							/>
-							<Results currentTopics={this.state.currentTopics} toggleTopic={this.toggleTopic} />
-						</div>
-					</div>
-				</ReactiveBase>
-			</section>
-		);
-	}
-}
+  return (
+    <section className="container">
+      <ReactiveBase
+        app="gitxplore-app"
+        url="https://xe6N9nDRV:51ea7a8a-6354-4b5f-83e1-12dce3b7ec47@arc-cluster-appbase-demo-ps1pgt.searchbase.io"
+        enableAppbase
+        theme={theme}
+      >
+        <div className="flex row-reverse app-container">
+          <Header currentTopics={currentTopics} setTopics={setTopics} />
+          <div className="results-container">
+            <DataSearch
+              componentId="repo"
+              filterLabel="Search"
+              dataField={[
+                "name",
+                "description",
+                "name.keyword",
+                "fullname",
+                "owner",
+                "topics"
+              ]}
+              placeholder="Search Repos"
+              iconPosition="left"
+              autosuggest={false}
+              URLParams
+              className="data-search-container results-container"
+              innerClass={{
+                input: "search-input"
+              }}
+            />
+            <Results currentTopics={currentTopics} toggleTopic={toggleTopic} />
+          </div>
+        </div>
+      </ReactiveBase>
+    </section>
+  );
+};
 
 export default App;
